requirements.txt
# Basic PyTorch and Hydra dependencies
torch>=1.9.0 # Specify a version that supports your desired features
torchvision>=0.10.0 # If you plan to use torchvision datasets/transforms
torchaudio>=0.9.0 # If you plan to use torchaudio

hydra-core>=1.1.0 # Core Hydra library
omegaconf>=2.1.0 # For structured configurations

# Optional, but common:
# tensorboard # For logging and visualization
# matplotlib # For plotting
# pandas # For data manipulation
# scikit-learn # For ML utilities
# jupyterlab # For notebooks
# pre-commit # For git hooks
